requirements.txt
click
pkg_resources
pyyaml
